// src/typings/TitleTeam.ts
// Copyright (c) 2022 Maxim "maxqwars" Maximenko <maxqwars@gmail.com>
//
// This software is released under the MIT License.
// https://opensource.org/licenses/MIT

export type TitleTeam = {
  decor: string[] | null;
  editing: string[] | null;
  timing: string[] | null;
  translator: string[] | null;
  voice: string[] | null;
};