// src/typings/TitlePosters.ts
// Copyright (c) 2022 Maxim "maxqwars" Maximenko <maxqwars@gmail.com>
//
// This software is released under the MIT License.
// https://opensource.org/licenses/MIT

import { Poster } from './Poster';

export type TitlePosters = {
  small: Poster | null;
  medium: Poster | null;
  original: Poster | null;
};